gateway: move discovery resolver into its own file
This branch moves the `Gateway::discover` method from
`gateway/src/server.rs` into its own file. This is to prepare for future
changes that will add more discovery-related logic to the gateway proxy,
which made sense to put in its own module (see #2333 for details). In
addition, I renamed the `Gateway::discover` method to
`Gateway::resolver`, for consistency with the similar method on
`Outbound`.

No functional changes.
